requirements.txt
pyqt5-tools
urllib3
nltk
bs4
inscriptis
validators
PyMySQL
